arch-arm: DPRINTF which XML we are sending to GDB
For some reason which I am still puzzled over, this function can return
the wrong XML descriptor (i.e. aarch64 in a 32-bit context); to aid in
tracing what's going on, I have added DPRINTFs to easily see which XML
we are returning.  Surprisingly, lifting the ternary to if, seems to
also result in returning the correct value.

Change-Id: I9d5e8df7c4e04759866f2abadfebdae814a94896